+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/histogram/index.spec.js
